// src/domains/shop/enums/shop-item-subtype.enum.ts
export enum ShopItemSubtypeE {
  GUN = 'Gun',
  CLOTHES = 'Clothes',
  SERVICES = 'Services',
  RESOURCES = 'Resources',
  EXPLOSIVE = 'Explosive',
  COMPONENTS = 'Components',
  AMMO = 'Ammo',
  SKIN = 'Skin',
  ELECTRICAL = 'Electrical',
  TOOL = 'Tool',
  TEA = 'Tea',
  CARD = 'Card',
  MEDS = 'Meds',
  ITEM = 'Item',
  BLUEPRINT = 'Blueprint',
  SKINS = 'Skins',
}